requirements.txt
Django==2.0.3
coverage==4.5.1
dj-database-url==0.5.0
django-allauth==0.35.0
django-braces==1.12.0
django-crispy-forms==1.7.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.0
django-floppyforms==1.7.0
django-imagekit==4.0.2
django-suit==0.2.26
dj-static==0.0.6
psycopg2==2.7.4
python-dateutil==2.6.1
model-mommy==1.5.1
django-ckeditor==5.4.0
pillow==5.0.0
gunicorn==19.7.1
static3==0.7.0
